angularfireDatabase + add-remove serv change
Done:
Implemented AngularFireDatabase on Diplsy service
Changed impementation of add-remove service functions to remove the use of tombstones.
adding stores the item with the itemBarcode as the parentFolder of the item. Makes delete method easier.

Do:
Alter implementation of add-remove service functions similar to lab 9, must find a way to delete the specified item by getting the parentFolder.