Fixed being able to bypass the sign-in page by entering a URL
Changed the routing to check that the user is signed in before displaying a page. If they are not signed in auth.guard will redirect to the sign-in page. Modified auth.service to save the logged in state as a cookie so that it will persist across the site. Modified the sign-out layout to change the cookie's logged in state to false.